# Testing Complete with Contacts Debugging # Language Switching through Bottom Sheet working

+++ app/src/main/java/com/karan/nishtharedefined/db/dao/ContactsDao.kt
@@ -15,7 +15,7 @@ interface ContactsDao {
     fun insertAllContacts(vararg contacts : Contact)
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insertContact(contact : Contact):Int
+    fun insertContact(contact : Contact): Long
 
     @Query("DELETE FROM Contact")
     fun deleteAllContacts()
